+++ src/routers/wrappers.ts
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ *
+ * This program and the accompanying materials are made available under the terms of
+ * the GNU Affero General Public License v3.0. You should have received a copy of the
+ * GNU Affero General Public License along with this program.
+ *  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
+ * SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
+ * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+import { RequestHandler } from 'express';
+import egoRequestWrapper from '../external/ego';
+
+/**
+ * Decorator to handle errors from async express route handlers
+ */
+export const wrapAsync = <PathParams, ResponseBody, RequestBody, Query, Locals extends Record<string, any>>(
+	fn: RequestHandler<PathParams, ResponseBody, RequestBody, Query, Locals>,
+): RequestHandler<PathParams, ResponseBody, RequestBody, Query, Locals> => {
+	return (req, res, next) => {
+		const routePromise: any = fn(req, res, next);
+		if (routePromise.catch) {
+			routePromise.catch(next);
+		}
+	};
+};
+
+export const wrapAsyncAuth = process.env.AUTH_ENABLED === 'true' ? egoRequestWrapper() : wrapAsync;
